[functions] rewrite of FunctionProvider to reduce code duplication
There is finally only one FunctionProvider, all specializations are gone.
The price for this is one additional line in all functions that should be
provided which is acceptable. See the added documentation of
FunctionProvider on how to proceed for new functions. The only drawback is
that there are still four places in functions.hh that need to be touched.
This is left as an exercise for someone who fancies static loops over tuple
types (not a tuple object), I am too tired now... :)