[container] fix deadlock
The problem was a locked method calling backend(), thus calling
ensure_uniqueness(), which in turn waited for a lock. The solution is to
first ensure uniqueness and obtain, then lock, then use a reference on the
unique backend. This is in general acchieved by
```
auto& backend_ref = backend(); // calles ensure_uniqueness()
std::lock...
// work with backend_ref
```

To make it a bit more explicit what is going on we could also let
ensure_uniqueness() return a `BackendType&` and use that instead of
backend().